started adding dynamic zoom display
Only have max and min zoom levels currently, but will use same conventions. Had to put all geojson data together in one file for it to work, so it's building off of the ipython notebook, which added lat and lng extent. That's what I'm using to determine whether a set of polygons should be displayed or not.