now preserves map state between page loads
This happens via cookies, and a rather elaborate process to store the on/off state of each chart as compactly as I can make it without encoding bytes in characters or something. Basically, since I want to keep track of all of the on/off states for all of the charts, I'm making a big binary number with as many digits as there are charts. Then, I make the number more compact by transforming it from a binary number to a hex number (base 2 to base 16). That's how it's stored in the cookie, and how it will be stored in the HTTP GET, so that these states can be bookmarked. If it weren't for that, I'd just make a huge cookie.