obsoleted "headers" php files
The set of 3 "headers" php files used to read the atlas metadata csv file and wrap that content in html have been made obsolete. The metadata is now an importable object in the modules, and the `sidebar_skeleton.js` module is in charge of creating that content on the page. The static list still needs to use content from the geojson file to populate the comprehensive list, but things are moving away from needing php files.